requirements.txt
antlr4-python3-runtime==4.8
llvmlite
requests